Separar entidades en archivos
