+++ SaludTechDeLosAlpes/Ingestion/Ingestion.Infraestructura/Persistencia/Entidades/ContextoProcesalEntity.cs
@@ -1,0 +1,7 @@
+namespace Ingestion.Infraestructura.Persistencia.Entidades;
+
+public class ContextoProcesalEntity
+{
+	public Guid Id { get; set; }
+	public string Etapa { get; set; }
+}
