+++ SaludTechDeLosAlpes/Ingestion/Ingestion.Infraestructura/Persistencia/Entidades/EntornoClinicoEntity.cs
@@ -1,0 +1,7 @@
+namespace Ingestion.Infraestructura.Persistencia.Entidades;
+
+public class EntornoClinicoEntity
+{
+	public Guid Id { get; set; }
+	public string TipoAmbiente { get; set; }
+}
