+++ SaludTechDeLosAlpes/Ingestion/Ingestion.Infraestructura/Persistencia/Entidades/ModalidadEntity.cs
@@ -1,0 +1,8 @@
+namespace Ingestion.Infraestructura.Persistencia.Entidades;
+
+public class ModalidadEntity
+{
+	public Guid Id { get; set; }
+	public string Nombre { get; set; }
+	public string Descripcion { get; set; }
+}
