+++ SaludTechDeLosAlpes/Ingestion/Ingestion.Infraestructura/Persistencia/Entidades/PatologiaEntity.cs
@@ -1,0 +1,7 @@
+namespace Ingestion.Infraestructura.Persistencia.Entidades;
+
+public class PatologiaEntity
+{
+	public Guid Id { get; set; }
+	public string Descripcion { get; set; }
+}
