+++ SaludTechDeLosAlpes/Ingestion/Ingestion.Infraestructura/Persistencia/Entidades/RegionAnatomicaEntity.cs
@@ -1,0 +1,8 @@
+namespace Ingestion.Infraestructura.Persistencia.Entidades;
+
+public class RegionAnatomicaEntity
+{
+	public Guid Id { get; set; }
+	public string Nombre { get; set; }
+	public string Descripcion { get; set; }
+}
