+++ SaludTechDeLosAlpes/Ingestion/Ingestion.Infraestructura/Persistencia/Entidades/SintomaEntity.cs
@@ -1,0 +1,8 @@
+namespace Ingestion.Infraestructura.Persistencia.Entidades;
+
+public class SintomaEntity
+{
+	public Guid Id { get; set; }
+	public string Descripcion { get; set; }
+	public ICollection<MetadatosEntity> Metadatos { get; set; } = new List<MetadatosEntity>();
+}
